+++ scripts/install.js
@@ -16,7 +16,6 @@ const pathPinturaJS = path.join(
     'pintura-iife.js'
 );
 const contentsPinturaJS = fs.readFileSync(pathPinturaJS, { encoding: 'utf8' });
-
 const pathPinturaCSS = path.join(projectRoot, 'node_modules', '@pqina', 'pintura', 'pintura.css');
 const contentsPinturaCSS = fs.readFileSync(pathPinturaCSS, { encoding: 'utf8' });
 
@@ -43,6 +42,24 @@ const contentsPinturaVideoCSS = fs.existsSync(pathPinturaVideoCSS)
     ? fs.readFileSync(pathPinturaVideoCSS, { encoding: 'utf8' })
     : undefined;
 
+// read optional dependencies for video encoding
+const pathWebMMuxer = path.join(
+    projectRoot,
+    'node_modules',
+    'webm-muxer',
+    'build',
+    'webm-muxer.js'
+);
+const contentsWebMMuxer = fs.existsSync(pathWebMMuxer)
+    ? fs.readFileSync(pathWebMMuxer, { encoding: 'utf8' })
+    : undefined;
+
+// read optional dependencies for video encoding
+const pathMp4Muxer = path.join(projectRoot, 'node_modules', 'mp4-muxer', 'build', 'mp4-muxer.js');
+const contentsMp4Muxer = fs.existsSync(pathMp4Muxer)
+    ? fs.readFileSync(pathMp4Muxer, { encoding: 'utf8' })
+    : undefined;
+
 // read component file
 const dest = path.join('bin', src);
 
@@ -57,26 +74,82 @@ data = data.replace('/*__PINTURA_JS__*/', () => contentsPinturaJS);
 
 // if pintura video extension installed
 if (contentsPinturaVideoJS) {
-    // inject webm duration fix so we don't get videos without duration on android (chrome bug)
-    const contentsWebmDurationFix = fs.readFileSync(
-        path.join('scripts', 'webm-duration-fix.min.js'),
-        {
-            encoding: 'utf8',
-        }
-    );
-    data = data.replace('/*__WEBM_DURATION_FIX_JS__*/', () => contentsWebmDurationFix);
-    data = data.replace(
-        '/*__WEBM_DURATION_FIX_APPLY__*/',
-        () => `
-// if is webm (android + chrome)
-if (value.type === 'video/webm') {
-    fixWebmDuration(value).then((blob) => {
-        // turn fixed blob into data uri
-        toDateURI(blob).then((str) => resolve([key, str]));
-    });
-    return;
-}`
-    );
+    // only if webm or mp4 muxer installed
+    let muxerVideoWriterDefault = '';
+    let muxerVideoWriter = '';
+
+    // media stream writer
+    const mediaStreamVideoWriterDefault = `createDefaultVideoWriter({
+    encoder: createMediaStreamEncoder({
+        imageStateToCanvas,
+    })
+})`;
+
+    const mediaStreamVideoWriter = `createDefaultVideoWriter({
+    ...(msg.editorOptions.videoWriter || {}),
+    encoder: createMediaStreamEncoder({
+        ...(msg.editorOptions.videoWriter ? msg.editorOptions.videoWriter.mediaStreamEncoder || {} : {}),
+        imageStateToCanvas,
+    }),
+})
+`;
+
+    // if webm muxer installed
+    if (contentsWebMMuxer) {
+        // replace webm muxer
+        data = data.replace('/*__VIDEO_MUXER_JS__*/', () => contentsWebMMuxer);
+
+        // default muxer options
+        muxerVideoWriterDefault = `
+createDefaultVideoWriter({
+    encoder: createMuxerEncoder({
+        muxer: WebMMuxer,
+        mimeType: 'video/webm',
+        imageStateToCanvas,
+    })
+}),
+`;
+
+        // custom muxer options
+        muxerVideoWriter = `
+createDefaultVideoWriter({
+    ...(msg.editorOptions.videoWriter || {}),
+    encoder: createMuxerEncoder({
+        ...(msg.editorOptions.videoWriter ? msg.editorOptions.videoWriter.muxerEncoder || {} : {}),
+        muxer: WebMMuxer,
+        mimeType: 'video/webm',
+        imageStateToCanvas,
+    })
+}),
+`;
+    } else if (contentsMp4Muxer) {
+        // replace webm muxer
+        data = data.replace('/*__VIDEO_MUXER_JS__*/', () => contentsMp4Muxer);
+
+        // default muxer options
+        muxerVideoWriterDefault = `
+createDefaultVideoWriter({
+    encoder: createMuxerEncoder({
+        muxer: Mp4Muxer,
+        mimeType: 'video/mp4',
+        imageStateToCanvas,
+    })
+}),
+`;
+
+        // custom muxer options
+        muxerVideoWriter = `
+createDefaultVideoWriter({
+    ...(msg.editorOptions.videoWriter || {}),
+    encoder: createMuxerEncoder({
+        ...(msg.editorOptions.videoWriter ? msg.editorOptions.videoWriter.muxerEncoder || {} : {}),
+        muxer: Mp4Muxer,
+        mimeType: 'video/mp4',
+        imageStateToCanvas,
+    })
+}),
+`;
+    }
 
     // inject scripts and styles
     data = data.replace('/*__PINTURA_VIDEO_CSS__*/', () => contentsPinturaVideoCSS);
@@ -89,6 +162,7 @@ if (value.type === 'video/webm') {
             () => `
 const {
     createDefaultVideoWriter,
+    createMuxerEncoder,
     createMediaStreamEncoder,
     plugin_trim,
     plugin_trim_locale_en_gb,
@@ -108,14 +182,8 @@ imageWriter: createDefaultMediaWriter(
     [
         // For handling images
         createDefaultImageWriter(),
-
-        // For handling videos
-        createDefaultVideoWriter({
-            // Encoder to use
-            encoder: createMediaStreamEncoder({
-                imageStateToCanvas,
-            }),
-        }),
+        ${muxerVideoWriterDefault}
+        ${mediaStreamVideoWriterDefault}
     ]
 ),
 `
@@ -124,15 +192,14 @@ imageWriter: createDefaultMediaWriter(
             '/*__PINTURA_VIDEO_WRITER__*/',
             () => `
 if (msg.editorOptions.imageWriter || msg.editorOptions.videoWriter) {
-    msg.editorOptions.imageWriter = createDefaultMediaWriter(msg.editorOptions.imageWriter || {}, [
-        createDefaultImageWriter(),
-        createDefaultVideoWriter({
-            ...(msg.editorOptions.videoWriter || {}),
-            encoder: createMediaStreamEncoder({
-                imageStateToCanvas,
-            }),
-        }),
-    ]);
+    msg.editorOptions.imageWriter = createDefaultMediaWriter(
+        msg.editorOptions.imageWriter || {},
+        [
+            createDefaultImageWriter(),
+            ${muxerVideoWriter}
+            ${mediaStreamVideoWriter}
+        ]
+    );
 }
 `
         );